# backend/other/inspiral-swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a server of InSpiral webpage.",
    "version": "1.0.0",
    "title": "InSpiral App",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "morvai.daniel96@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "todo.swagger.io",
  "basePath": "/v2",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/trainer/login": {
      "post": {
        "tags": [
          "trainer"
        ],
        "summary": "Log in trainer",
        "description": "",
        "operationId": "addTrainer",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "description": "Trainer's email to login",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "query",
            "description": "The password for login",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Login OK",
            "schema": {
              "type": "string"
            },
            "headers": {
              "JWT token": {
                "type": "string",
                "description": "token of trainer returned"
              }
            }
          },
          "400": {
            "description": "Invalid email/password supplied"
          }
        }
      }
    },
...
    "/admin/trainer": {
      "get": {
        "tags": [
          "trainer"
        ],
        "summary": "Gets logged in trainer",
        "description": "",
        "operationId": "getTrainer",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Trainer"
            }
          },
          "404": {
            "description": "Trainer not found"
          }
        }
      },
      "put": {
        "tags": [
          "trainer"
        ],
        "summary": "Update trainer",
        "description": "",
        "operationId": "editTrainer",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Trainer to be updated",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Trainer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Update OK"
          },
          "404": {
            "description": "Trainer not found"
          }
        }
      }
    },
    "/admin/poll": {
      "post": {
        "tags": [
          "poll"
        ],
        "summary": "Add a new poll",
        "description": "",
        "operationId": "addPoll",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Poll to be added",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Poll"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Poll created"
          }
        }
      },
      "get": {
        "tags": [
          "poll"
        ],
        "summary": "Get trainer's polls",
        "description": "",
        "operationId": "getTrainerPoll",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Poll"
              }
            }
          }
        }
      }
    },
    "/admin/poll/{pollId}": {
      "put": {
        "tags": [
          "poll"
        ],
        "summary": "Update trainer's poll",
        "description": "",
        "operationId": "updateTrainersPoll",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "pollId",
            "in": "path",
            "description": "ID of poll to be updated",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Poll to be updated",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Poll"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Update OK"
          },
          "404": {
            "description": "Poll not found"
          }
        }
      },
      "delete": {
        "tags": [
          "poll"
        ],
        "summary": "Delete trainer's poll",
        "description": "",
        "operationId": "deleteTrainersPoll",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "pollId",
            "in": "path",
            "description": "ID of poll to be deleted",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Delete OK"
          },
          "404": {
            "description": "Poll not found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Trainer": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "image": {
          "type": "string"
        }
      }
    },
...
    "Poll": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "trainerId": {
          "type": "integer",
          "format": "int64"
        },
        "trainerName": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "date": {
          "type": "string"
        }
      }
    },
...
  },
...
}